rename authoring hook state file
