+++ packages/editor/src/functions/addMediaNode.ts
@@ -33,7 +33,7 @@ import { Entity, EntityUUID } from '@etherealengine/ecs/src/Entity'
 import { defineQuery } from '@etherealengine/ecs/src/QueryFunctions'
 import { AssetLoaderState } from '@etherealengine/engine/src/assets/state/AssetLoaderState'
 import { PositionalAudioComponent } from '@etherealengine/engine/src/audio/components/PositionalAudioComponent'
-import { addAuthoringHook } from '@etherealengine/engine/src/gltf/AuthoringHookSystem'
+import { addAuthoringHook } from '@etherealengine/engine/src/gltf/AuthoringHookState'
 import { ImageComponent } from '@etherealengine/engine/src/scene/components/ImageComponent'
 import { MediaComponent } from '@etherealengine/engine/src/scene/components/MediaComponent'
 import { ModelComponent } from '@etherealengine/engine/src/scene/components/ModelComponent'
