+++ packages/engine/src/gltf/AuthoringHookState.tsx
